optimize test assertions - minimize work if check succeeds
Some tests contain huge numbers of assertions (~8 billion assertions).
The timezone stuff is so fast that for these tests, the ASSERT_EQ
machinery ends up taking a majority of the test time.

This modifies ASSERT_EQ and ASSERT_EQ_QUIET so that these do as little
work as possible on success